Factor out parsing and stringifying key input
