+++ ts/key-input-registerer.ts
@@ -1,4 +1,4 @@
-import { KeyInputElement } from "./key-input.js";
+import { KeyInputElement } from "./key-input-elements.js";
 
 export function register() {
   KeyInputElement.register();
